gcp-document-apis/functions/ocr_processor/requirements.txt
google-cloud-vision==3.4.4
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
